relocating blocks position
